# openapi.yaml
openapi: 3.0.3
info:
  title: HOLY BIBLE API
  version: 0.1.0
  description: Backend API for Bible content, search, annotations, and admin import.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
paths:
...
  /v1/versions:
    get:
      summary: List bible versions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BibleVersion'
    post:
      summary: Create new bible version (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BibleVersionCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BibleVersion'
  /v1/versions/{versionCode}/books:
    get:
      summary: List books in a version
      parameters:
        - in: path
          name: versionCode
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
  /v1/versions/{versionCode}/books/{bookId}/chapters/{chapterNumber}:
    get:
      summary: Get chapter with verses
      parameters:
        - in: path
          name: versionCode
          required: true
          schema:
            type: string
        - in: path
          name: bookId
          required: true
          schema:
            type: string
        - in: path
          name: chapterNumber
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterWithVerses'
  /v1/versions/{versionCode}/books/{bookId}/chapters/{chapterNumber}/verses/{verseNumber}:
    get:
      summary: Get single verse
      parameters:
        - in: path
          name: versionCode
          required: true
          schema:
            type: string
        - in: path
          name: bookId
          required: true
          schema:
            type: string
        - in: path
          name: chapterNumber
          required: true
          schema:
            type: integer
        - in: path
          name: verseNumber
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Verse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BibleVersion:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        title:
          type: string
        language:
          type: string
        description:
          type: string
        licenseId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        importedAt:
          type: string
          format: date-time
          nullable: true
    BibleVersionCreate:
      type: object
      required: [code, title, language]
      properties:
        code:
          type: string
        title:
          type: string
        language:
          type: string
        description:
          type: string
        licenseId:
          type: string
    Book:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        order:
          type: integer
    ChapterWithVerses:
      type: object
      properties:
        id:
          type: string
        number:
          type: integer
        verses:
          type: array
          items:
            $ref: '#/components/schemas/Verse'
    Verse:
      type: object
      properties:
        id:
          type: string
        number:
          type: integer
        text:
          type: string
        checksum:
          type: string
...